+++ src/hpk.rs
@@ -14,19 +14,25 @@ pub struct QUICHeaderProtection {
 #[pymethods]
 impl QUICHeaderProtection {
     #[new]
-    pub fn py_new(key: &PyBytes, algorithm: u16) -> Self {
-        QUICHeaderProtection {
-            hpk: HeaderProtectionKey::new(
-                match algorithm {
-                    128 => &AES_128,
-                    256 => &AES_256,
-                    20 => &CHACHA20,
-                    _ => panic!("unsupported"),
-                },
-                key.as_bytes(),
-            )
-            .expect("FAILURE"),
-        }
+    pub fn py_new(key: &PyBytes, algorithm: u16) -> PyResult<Self> {
+        let inner_hpk = match HeaderProtectionKey::new(
+            match algorithm {
+                128 => &AES_128,
+                256 => &AES_256,
+                20 => &CHACHA20,
+                _ => return Err(CryptoError::new_err("Algorithm not supported")),
+            },
+            key.as_bytes(),
+        ) {
+            Ok(hpk) => hpk,
+            Err(_) => {
+                return Err(CryptoError::new_err(
+                    "Given key is not valid for chosen algorithm",
+                ))
+            }
+        };
+
+        Ok(QUICHeaderProtection { hpk: inner_hpk })
     }
 
     pub fn mask<'a>(&self, py: Python<'a>, sample: &PyBytes) -> PyResult<&'a PyBytes> {